Better version for multi
- I was attaching after break-inserts and symbol-list-lines, which was
confusing cpptools and which resulted in a exec-run and finally gdb-exit.
In gdbserver --attach, target-select remote results in an already attached
state, so I changed target-select remote to be target-select extended-remote
and target-attach, which results an equivalent state. Now cpptools is no longer
confused and does not attempt to kill the gdb session.